Organize code.
