+++ frontend/src/containers/App.js
@@ -5,7 +5,7 @@ import "../styles/App.scss";
 import UniversalLandingPage from "../containers/UniversalLandingPage.js";
 import HostMenuView from "./HostMenuView";
 import DecksView from "./DecksView";
-import HostRoomView from "./HostRoomView";
+import HostGamePage from "./HostGamePage";
 
 import { Route, Switch } from "react-router-dom";
 
@@ -17,7 +17,7 @@ function App() {
 
         <Route path="/host/edit/:id" />
         <Route path="/host/decks" component={DecksView}/>
-        <Route path="/host/room" component={HostRoomView} />
+        <Route path="/host/room" component={HostGamePage} />
         <Route path="/host" component={HostMenuView} />
 
         <Route path="/" component={UniversalLandingPage} />
