+++ frontend/src/containers/HostGamePage/components/PlayerList/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from '../styles/PlayerList.module.scss';
+import styles from './styles.module.scss';
 
 export default function PlayerList({ players = [] }) {
   return (
