+++ frontend/src/containers/HostGamePage/components/PlayerList/styles.module.scss
@@ -1,5 +1,5 @@
-@import 'colors';
-@import 'scroller';
+@import '../../../../styles/colors';
+@import '../../../../styles/scroller';
 
 .container {
   width: 55vw;
